refactor(ui-document): split render_context_menu by menu section
CODE-4: extract the ~1065-line render_context_menu into per-section
builders in context_menu/sections.rs (copy/cell, row, filter, order,
export, selection, document/preview). render_context_menu now composes
them. Pure extraction; identical menu structure, order and listeners.
